requirements.txt
fastapi
uvicorn
python-multipart
watchdog
numpy
pandas
timm
pillow
torchvision
faiss-cpu
